Split up NumeroteMigrator class.
